refactor: change structure of the project (#7)
