+++ src/index.html
@@ -122,8 +122,6 @@
       </div>
     </div>
   </footer>
-  <!--<script type="module" src="script.js"></script>
-  <script type="module" src="modal.js"></script>-->
 </body>
 
 </html>
