+++ src/index.ts
@@ -1,11 +1,9 @@
 import './style.scss';
-import './script.ts';
-import './modal.ts';
-import './goods.ts';
-import './helpers/get-values';
-import './helpers/filters';
-import './helpers/sorting';
-import './pages/product-details/details.scss';
+import './pages/product-catalog/script.ts';
+import './modules/goods.ts';
+import './modules/get-values';
+import './modules/filters';
+import './modules/sorting';
 import './pages/product-details/details';
 import './pages/cart/cart';
-import './components/Header/Header';
+import './components/Header';
