+++ src/pages/404/404.html
@@ -63,7 +63,6 @@
             </div>
         </div>
     </footer>
-    <!-- <script type="module" src="script.js"></script> -->
 </body>
 
 </html>
